requirements.txt
asgiref==3.6.0
attrs==22.2.0
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.1.7
djangorestframework==3.14.0
idna==3.4
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.1.1
pluggy==1.0.0
psycopg2-binary==2.9.5
pytest==7.2.2
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
Random-Word==1.0.11
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.15
